Rearrange resources
