+++ omise/resources/account.py
@@ -1,5 +1,4 @@
-from omise import Base, _as_object
-from .main_resource import _MainResource
+from omise import Base, _as_object, _MainResource
 
 
 class Account(_MainResource, Base):
