+++ omise/resources/charge.py
@@ -1,7 +1,6 @@
 import copy
 
-from omise import Base, _as_object, LazyCollection
-from .main_resource import _MainResource
+from omise import Base, _as_object, LazyCollection, _MainResource
 
 
 class Charge(_MainResource, Base):
