+++ omise/resources/source.py
@@ -1,5 +1,5 @@
-from omise import Base, _as_object
-from .main_resource import _MainResource
+from omise import Base, _as_object, _MainResource
+
 
 class Source(_MainResource, Base):
     """API class for creating Source.
